requirements.txt
fastapi
uvicorn
openai-whisper
pydub
requests
python-multipart
